Add interface to run queries inside a transaction
Add a `transactional` method to both the pg.Pool and pg.Client classes.
This `transational` method takes a promise-returning function and runs
it inside a transaction.

If promise that results from that function resolves, the transaction
will be committed. If the promise that results from that function
rejects, the transaction will be rolled back.

Fixes #17